Add transpose parameter.
This parameter allows notes generated by the arpeggiator to be
transposed up or down by a given number of semitones.

This isn't in the Ardour version of the plugin (there are other MIDI
plugins which do that kind of thing), but is useful in the stand-alone
version to quickly adjust the note range for external groove boxes or
synths which do not offer such capabilities.