Allow selection of the raptor patch which receives mapped MIDI CCs.
By default, raptor listens to CC input on all channels and maps it to
parameters as determined by the MIDI mapping. This is also known as
"omni" CC map mode. If multiple raptor instances are running, this means
that MIDI mappings are applied to all the instances, so that parameters
in the different instances will change in lockstep.

If you only want to change the parameters of a single instance, you can
now click on the button in the upper left corner of the corresponding
panel, called the "omni" button. The button will then change to a blue
background color to indicate that this is now the only instance which
receives all MIDI parameter changes.

Only one raptor instance can be operated in non-omni CC map mode at any
one time. You can change the target raptor instance by just clicking on
the omni button in the corresponding panel.

Clicking the blue omni button again changes back to omni CC map mode
which makes all raptor instances listen for parameter changes again.