Add support for the Novation LaunchControl XL (experimental).
This enables some hard-wired MIDI bindings for the Novation
LaunchControl XL which makes it easy to switch the global ccmaster (the
target instance which receives MIDI-mapped controls if you're running
multiple raptor instances).

These bindings will only work if the LaunchControl is switched to the
first factory preset (which transmits on MIDI channel 9), and of course
the LaunchControl needs to be connected to Pd's first MIDI input. It
binds the Device Hold + Prev/Next Device Select and Device Hold + Device
Bank button combinations so that they will switch the ccmaster
accordingly.

For convenience, we have this enabled by default, which shouldn't
normally cause any issues. But you can disable this functionality if you
don't need it, by changing the value of the launchcontrol variable in
lib/raptor.pd_lua to 0.