Record meter and tempo information in loop files.
This means that we can now restore the original meter and tempo a loop
was recorded with when loading a loop file. (In Raptor 6, that
information was kept in the preset instead, but Raptor 7 doesn't do that
any more in order to not disrupt live performances. It makes much more
sense to keep that information with the recorded sequence anyway.)

Also, it is now possible to load and save loop files even if the prefix
of the loop filename does not denote an existing preset, by just typing
the prefix into the preset field in the panel.