Update the Launch Control XL map.
This uses the new and tidier output format. Also added the transport
rewind and pos/anacrusis functions which are useful as stutter and scrub
controls in live performance. Note that the raptor toggle was moved to
the send select "down" button so that we can assign a fader to the pos
control.