Add pgmset mode (experimental).
Normally, Raptor passes through program change messages, this new option
makes it possible to use program changes to switch presets instead. It
works like regular CC MIDI mapping in that multiple Raptor instances
will process the program changes in omni mode, but you can set the
ccmaster to have the messages directed at a specific Raptor instance.

This mode is still a bit experimental, and only documented inside the
patch for now. While it may be useful for automation purposes when
running Raptor as a plugin (i.e., using plugdata), it isn't enabled by
default. At present, the only way to enable pgmset mode is the toggle
you can find in the init subpatch of the main patch (which see).