Improved DJ controller support.
This is still experimental, but beginning to take shape.

- Implemented deck assignment (this is in the init subpatch), possible
  values are -1 (off), 0 (omni, this is the default), 1 (left deck), and
  2 (right deck).

- Implemented playback, cue (a.k.a. rewind to anacrusis), sync (reset
  playback position to 0), and of course the jog wheels now control the
  playback positions of each corresponding Raptor instance. (To make
  this work, selecting a deck in the init subpatch now automatically
  disables SPP synchronization in the time subpatch, so that each Raptor
  instance can use its own anacrusis when playback starts.) Note that
  these are all hard-wired controls, as they require custom
  functionality that isn't readily available for MIDI mapping (yet).

- MIDI mapping now works as expected for the two decks on an individual
  basis. The ccmaster setting is generally ignored for the deck-specific
  controls, since messages are filtered by the deck numbers instead.

- Global controls can be mapped as well, and will go to the Raptor
  instances for all decks by default. Pressing one of the shift keys
  with the global controls sends to the current ccmaster instead. The
  hard-wired BROWSE encoder (used to change presets) works in the same
  fashion.

The time and init subpatches were modified to implement the new
features. There's no real documentation yet, so you'll have to dive into
the source or play around with the jog wheels and the sync and cue
buttons to figure out how it all works. There's no ready-made MIDI
mapping for the Hercules DJControl yet either, but it should be a piece
of cake to get the tempo faders and loop and play buttons set up so that
they are bound to the corresponding Raptor controls.

Note that in contrast to standard DJ software, there's no way to just
start a single Raptor instance; once you press play on an instance, it
becomes the time master and all instances start rolling in perfect
sync. But of course you can easily mute individual Raptor instances, so
you might want to bind a deck control to that function.

What remains to be done is to implement the volume faders and the cross
faders on the DJControl, so that they do something sensible by default.
Maybe also do something with some of the pads (e.g., by mapping them to
the bars in a loop). And, of course, supply a ready-made MIDI map.