MIDI pass-through heeds ccmaster.
For convenience, check for ccmaster in order to direct messages to a
single Raptor instance in an ensemble. This is a simple change, but it
makes live so much easier if you want to change, e.g., pan and volume on
a single instrument. Before this, you had to fiddle around with the
inchan settings to make this work. Now you can just select the instance,
send the CC, and be done with it.