djcontrol: Add volume and balance controls.
These are mapped to the volume fader and the filter knob, respectively.

Note that Raptor is MIDI only, so both controls are implemented by
simply remapping the MIDI data to CC7 and CC8. They will thus depend on
the capabilities of your synthesizer, and are 7 bit values only, so
don't expect any miracles.

Since everything needs to be done in MIDI, there's no simple way to
implement the cross-fade control, so we left that one out for now. For
the time being, you can just employ the usual flipping of the volume
faders to achieve that effect.