dijcontrol: Add hot-cue pad controls.
This assigns the mode 1 (hot-cue) pads for both decks, both shifted and
unshifted, so that they do something useful in live performance.

- The unshifted pads 1-4 (or 1-8 depending on the device version) call
  up the corresponding bar in the playing loop with a smooth transition
  (i.e., the playback position within the bar remains unchanged).

- The shifted pads work the same, but switch to the beginning of the
  corresponding bar immediately (i.e., resetting the playback position
  to the beginning of the bar).

In any case, the change takes effect with the next pulse.

These functions are only available when a loop is currently playing.
Otherwise, the pads are unbound and can thus still be mapped using MIDI
learn if no loop is playing.