djcontrol: Give some feedback (PLAY status and "energy" LED).
You'll only get these if you connect Pd's second MIDI output to the DJ
Control device. Indicates playback status on the PLAY buttons. Also, the
"energy" led on the big encoder flashes in the rhythm, as determined by
the actual pulse weights and velocities produced by the arpeggiator.

NOTE: The led could get rather busy with complex meters if we'd simply
flash all pulses, so we only do the n most salient pulses, where the
value n is currently hard-coded (8, at present). You might want to
experiment with different values there, check the
raptor:djcontrol_pulse() method for details.

The velocity value is just taken as is to determine the color of the
led. This seems to work reasonably well with the color codes described
in Hercules' MIDI Commands manual for the DJ Control Inpulse 200.