djcontrol: Reworked SYNC once more.
The way we implemented it previously, as a variation of the CUE control,
it was pretty useless and didn't behave in a way that you'd expect it
to. Instead, it now syncs the playback position of all Raptor instances
to that of the time master, which seems to be the right thing. Please
check djcontrol.map for details.