djcontrol: Fixes for better compatibility with other Hercules devices.
Swapped the shifted und unshifted assignments of the VINYL/LOOP button.
The Inpulse 200 MK2 is the odd man out there, the new assignment
hopefully works better with the other devices. Please check the
(updated) djcontrol.map which, as usual, has all the details.

More feedback (IN/OUT is lit when a loop is active, CUE and SYNC buttons
are always on).

New mappings (the encoder above IN/OUT in the loop section can be used
to change loop size, pressing the encoder toggles loop mode, like
LOOP a.k.a. SHIFT+VINYL).

VINYL mode is now enabled by default, which seems convenient, and makes
it possible to do scratching on controllers which don't even have that
button.

Tested with the Inpulse 200 MK2 and the Inpulse 500.