djcontrol: Refactoring the feedback.
The djcontrol subpatch had become rather unwieldy, so we moved most of
the feedback logic into raptor.pd_lua now. While at it, we also added
feedback for the big 1 and 2 backlights on the Inpulse 500, so that they
light up as soon as you touch one of the jog wheels in vinyl mode (or if
transport is stopped).