Improved handling of ccmaster next/prev wrap-around.
We want to go back to omni (no ccmaster) before wrapping around.
Wrap-around will then happen when repeating the operation. This has the
advantage that switching off the ccmaster becomes possible with just the
next/prev operations, improving usability on devices which only have
these, like the Nektar PACER.

Actually, by a lucky accident this was already the behavior for the
case of a single Raptor instance. We now made it work that way for
multiple instances as well.