launchpad: Set the sensitivity of the pads in session view.
Only trigger the button if the velocity is at or above the threshold. A
positive velocity below the threshold just prints the current binding
and value (if any) in the console window instead.

Also, the exact behavior depends on the control that the pad is mapped
to. On toggled controls, it toggles, otherwise it works like a latching
fader which inputs a value depending on the velocity (if it is above the
threshold) and snaps back to 0 (or the middle position, in case of a
bipolar control) as soon as the pad is released.