Add a GUI for the device driver configuration.
Implemented in lib/config.pd. This is available in the init subpatch.
It makes it more convenient to configure various driver defaults which
could only be set in the source previously. (Check the comments near the
beginning of raptor.pd_lua for details.) The available options are
toggles to enable and disable individual device drivers, and some
additional parameters of the Launchpad and DJ Control drivers.

The patch has a Submit button to send the configuration data to Raptor.
The Defaults button reverts to the factory values. You can also save the
config patch with the current settings to make your changes permanent.

Also, the MIDI note and CC processing chain was optimized, bypassing
drivers if MIDI data comes from ports which they don't listen to.