Make the Launchpad fader bank orientation configurable in the source.
See the launchpad_fader_orientation variable near the beginning of
raptor.pd_lua. This isn't in the config patch yet, so you'll have to
edit it in the source code for now. It is a table with four entries, one
for each fader bank (Volume, Pan, Send A, Send B); 0 means vertical, 1
horizontal. The default is { 0, 1, 1, 1 }, meaning that the first fader
bank is vertical, all others horizontal, which better matches the layout
of the parameter panel in the patch.