Overhaul of the Launchpad grid color system.
The color coding of the pads on the grid in session mode was previously
done statically, by pad position. Thus the pads wouldn't move if the
MIDI mapping was changed. This has been fixed; the colors of the pads
are now assigned by mapped function, not position, which works much
better if you change the default pad bindings in launchpad.map.

Moreover, we can now show alternative colors for *all* toggles, not only
the "play" and "loop" buttons, which is another big plus of the new
system. (There are still no color gradients for the pads which represent
continuous or enumerated parameters, however. We might still tackle this
in the future, at least for the enumerated parameters.)

Also, fixed some bugs were feedback messages could be emitted by
multiple Raptor instances, by the wrong instance, or too early in the
intitialization sequence.