Launchkey: Make mode 1 (Volume) the default knob mode in session mode.
This matches the standalone bindings in the MIDI map (by function, not
by CC), which is a good thing. The Launchkey seems to pick mode 3 (Pan)
by default in session mode. Thus, if we don't set this, it could lead to
some nasty surprises if, when switching from standalone to session mode,
the knobs all of a sudden don't control the parameters you think they do.