Add a fader pickup mode, enabled by default.
If a CC value doesn't reflect the current value of the Raptor parameter
to which it is mapped, you'll have to adjust the control until it does
before the mapped value starts changing.

If you know any modern DAW then you're probably familiar with this. We
have this enabled by default, which is highly recommended, as it vastly
improves the user experience if you're using any MIDI controller with
knobs and faders mapped using Raptor's MIDI learn facility.

But you can still change this back to the previous behavior in the
config patch, if you prefer to have parameters values jump as soon as
you move a control. Note that this setting is global, so you can't
change it for an individual Raptor instance when running an ensemble.