Some compatibility fixes for the rest of the Launchkey MK3 lineup.
This is for the bigger Launchkeys (not the Mini). We also emulate the
Stop button which isn't a real Raptor function. On the Mini, you can now
get that function with Shift+Play.

This is based entirely on the programmer's reference, so this still
needs to be tested on a real device.