Add a fifth "Extra" bank of faders to the Launchpad driver.
This bank is identical to the Launchkey's Device bank, which has a
selection of some commonly used performance controls from the other
fader banks on it. The new bank can be accessed with the Clip Stop
button (the 5th scene launch button on Pro/X/Mini, or the last button in
the bottom row of the LP Pro).

Note that the previous assignment of "Stop Clip" as a note/drum toggle
in note mode was pretty much useless on the LP Mini and X anyway. (On
the X, we have the Note button for that, and the Mini has no dedicated
note mode at all, so you have to use the custom Keys and Drums modes
instead.) On the LP Pro we replaced "Stop Clip" with the "Clear" button,
which has the added benefit that this button is much closer to the Note
button. (Unfortunately, the Note button itself can't be used for that
purpose on the Pro, because it generates no MIDI data in note mode.)