Sign-on message on Launchpad and Launchkey.
Display the text "Raptor 7 ready" during startup to indicate that Raptor
is up and running on the device.

On the Launchkey, the text is shown on the LCD screen (not available on
the Mini). On the Launchpad X and Mini, the text scrolls across the grid
and the right-hand side buttons; at any time, you can press any of the
mode buttons at the top to stop the animation. (The scrolling text
display is not supported on the Launchpad Pro, but there it's easy to
tell that Raptor grabbed the device since the buttons will light up
around the grid in different colors.)

The messages to be shown are in the variables launchpad_welcome and
launchkey_welcome near the beginning of raptor.pd_lua. You can change
the messages with these variables. If you find the messages distracting,
you can also turn them off by setting the variables to nil instead of a
text message.