Launchkey: Some more refactoring and an improved instance display.
When a Raptor instance is selected, we now show the instance number
(numbering the instances from 1), without the internal id which doesn't
help the user. Instead we add the preset name set in that instance, to
help identify the instance.

Also fixed a display initialization issue, apparently caused by a race
condition on Linux/ALSA between the MIDI message which puts the
Launchkey into DAW mode, and subsequent sysex messages for initializing
the LCD display. We now make sure that the sysex messages are sent at a
later time, using an extra timer callback.