Turn self.master into a global time_master variable.
This is shared between all instances now. We also introduced a utility
function raptor:check_master() to check whether the current instance is
the time master. In particular, this helped to resolve a bug where the
state of the Launchpad and Launchkey pads in session mode wasn't updated
in some situations.

Morerover, we untangled the Launchpad and Launchkey ccmaster hand-over.
This is in two separate methods now, so that ccmaster switches in the
Launchpad and the Launchkey work independently from each other.
Previously, the hand-over wouldn't be invoked on the Launchkey if no
Launchpad device was connected.