Launchkey: Device Select and Device Lock buttons.
Device Select now brings up a strip with buttons to select a Raptor
instance at random, very similar to the selection buttons in the extra
row at the bottom of the Launchpad Pro MK3.

Device Lock doesn't actually lock anything, it just recalls the
information about the selected Raptor instance and preset to the LCD
display.

We also moved the CCs for the faders on the larger Launchkey models (49
and up), so that they don't interfere with our existing bindings.
Together with the "extra" buttons (Capture MIDI, Quantise, Click, Undo),
these are now in their own CC block 57-69 on channel 32 (channel 16 on
the 2nd Pd MIDI port) which currently isn't used for anything else
(fingers crossed).

Finally, we also fixed some finalization issues where the state of
Raptor instance buttons on the Launchpad, Launchkey, and DJ Control
devices would stick around when a Raptor instance is terminated.