Add a few more loop examples, all on slot #1.
The loops are intended to be used with the raptors.pd and raptors2.pd
sample ensembles, so that these can be demoed without live note input.

If you have a Launchpad or Launchkey, you can load the loops by pushing
the loop-next pad (the rightmost blue pad in the bottom row) in session
view once, then the loop-load pad (the teal pad in the row above). A
similar procedure works for the Launch Control XL and the MIDIMIX, using
the two button rows at the bottom. On the DJ Control, push the LOOP OUT
button once and then the LOAD button (for the raptors2.pd example,
you'll have to do that for each deck).