Add Launchkey identification, other cosmetic changes.
We need to distinguish the Launchkey 88 from the other MK3 devices since
it uses a different sysex id, which is needed to send data to the LCD
screen. Thus we now send a device inquiry on port #2 during startup and
pick up the results later, similar to the Launchpad. In contrast to the
Launchpad, we don't depend on this; by default we just assume id 15,
which should work with most of the MK3 controllers.