Rework the Launchkey drum grid to make it scrollable.
The drum grid can now be scrolled up and down with the up and down arrow
keys in Drum pad mode, and the pads now show the same colors for the
four 4x4 grids as on the Launchpad drum grid. Moreover, a darker color
is used for the right 2x8 group which would be above the first 2x8 in
the Launchpad layout, so that the layout becomes more apparent.