Add two drum grids below the existing four on the Launchkey.
This extends the range of the drum grids available on the Launchkey to
include the note ranges 4-19 and 20-35 below the four 4x4 grids in the
Launchpad drum mode which starts at note 36. Note that the new grids are
mostly below the standard GM drumkit range, but may offer some useful
sounds with some synth and drumbox patches.

Also added a display of the drum grid range on the LCD screen, which
gets shown when switching to Drum mode or when scrolling the drum grid.