Added missing initialization of the arpeggiator's raptor member.
This went unnoticed for a really long time, because we only test this
value to be nonzero, which yields true with a nil value. However, the
rest of the external actually assumes this value to be zero at startup,
which caused Raptor to produce raptor-mode arpeggios by default even
though the raptor toggle in the panel was off.

At long last, this is fixed now.