Fix the pad mode feedback for the Launchkey Mini.
The Launchkey Mini doesn't have any of the chord and device select modes
of the larger Launchkey models. If we try to set these (via feedback) on
the Mini, it forces all connected Launchkeys back to drum mode. We
prevent this now.

Note that we still allow all of the custom pad modes which *are* present
on the Mini (at least with the latest firmware), even though only the
first of these can be selected on the device itself.

Also note that we don't need this workaround for the knob a.k.a. pot
modes, because the Mini has all the knob modes of the larger models.