Add a metronome click for the Launchkey.
This can be turned on or off with the Click button on the larger
Launchkey models (LK 25 and up). The implementation is rather simplistic
right now, it just sends out notes on channel 10 of the first output
port, so it needs a synth with a GM-compatible drum channel to work.

The note number and volume can be set with some configuration variables
in the Launchkey config section at the beginning of raptor.pd_lua.

Also, the launchpad_n_pulses config variable is used to determine the
number of pulses per bar to play, which means that if you have both a
Launchpad and a Launchkey hooked up, the Launchpad's pulse display and
the Launchkey's metronome clock will run in sync. Also note that this
variable can be set in the config subpatch, so you can configure it at
runtime.