Turn the metronome click into a MIDI-mappable Raptor function.
This moves the metronome click from the Launchkey driver into Raptor
itself. It also adds a corresponding toggle in the main patch which can
be MIDI-mapped, so that this feature becomes generally available through
MIDI learn no matter which controller you use (if any).

A binding for the Click button has been added to the Launchkey map, so
that this continues to function as before.