Device select mode emulation for the Launchkey Mini.
Implemented a facsimile of device select mode which also works on the
Mini. This uses the Stop/Solo/Mute button along with the bottom pad row
in session mode to give a similar experience to what the Device Select
button does on the larger Launchkey models.

This seems way more useful than the standalone bindings for the > and
Stop/Solo/Mute buttons which have both been removed from DAW mode (note
that the latch and mute toggles are also available on the launch grid).