Launchpad Pro: Rework the feedback of the track select buttons.
Previously the entire strip of track select buttons would light up in
blue no matter how many Raptor instances there actually were, which was
a little confusing. Now the buttons only light up if there's an instance
which can be selected, so that you can see at a glance how many
instances there are. This is also how the device and track select
buttons work on the Launchkey.