CC16 on channel 1 selects preset (experimental).
Added another hard-wired mapping for selecting presets. CC16 0-19
selects preset 1-20 from the quick dial. I mainly use this to bind one
of the Q-links on an AKAI MPC for convenience, but this will work with
any MIDI gear that can be programmed to send CC16 (general purpose
slider 1) on channel 1.