Change pgmset to a 3-way switch (1 enable, 0 disable, -1 pass-through).
This used to be a 2-way switch where the default 0 value always passed
through program change messages to the output, but it turns out that
this gets in the way if we want to add custom functionality which uses
program changes in other, more creative ways. So we now just ignore
program change messages if the option is disabled, and require an
explicit -1 (or other negative) value to enable pass-through.

As before, enabling the option (1 or other positive value) means that
program change messages are used for direct preset selection.