Add a new pc abstraction to customize preset selection.
This is to be invoked in the main raptor patch as

  pc $0 n1 n2 ... n8

where $0 is the $0 of the main patch, and n1, n2, ..., n8 are the
numbers of presets to be selected with PC1, PC2, ..., PC8, respectively.

Note that this only works if you have direct preset selection turned off
in the init subpatch. It gives you quick access to (up to) eight
different presets in a live performance using, e.g., buttons or foot
switches which output program change messages on a MIDI device connected
to MIDI input port #1.

The raptor7-pc1.pd patch is a variation of raptor7.pd which has an
instance of the pc abstraction to demonstrate the feature.